packages/currency-quote/currency_quote-3.2.8-py3-none-any.whl/currency_quote/application/use_cases/get_last_currency_quote.py
# src/currency_quote/application/use_cases/validate_currency.py
from currency_quote.adapters.outbound.currency_api import CurrencyAPI
from currency_quote.domain.services.get_currency_quote import GetCurrencyQuoteService
from currency_quote.domain.entities.currency import CurrencyQuote


class GetLastCurrencyQuoteUseCase:
    @staticmethod
    def execute(currency_quote: CurrencyQuote) -> dict:
        quote_service = GetCurrencyQuoteService(
            currency=currency_quote, currency_repository=CurrencyAPI
        )
        return quote_service.last()
